fixing some compilation warnings, one more

+++ objVisualizer/core.hpp
@@ -533,7 +533,7 @@ inline std::ostream& operator<<( std::ostream& os, const face& p )
 inline std::ostream& operator<<( std::ostream& os, const std::vector<face>& p )
 {
 	os << std::endl;
-	for ( int i = 0; i < p.size( ); ++i )
+	for ( size_t i = 0; i < p.size( ); ++i )
 		os << "\t" << p[i] << std::endl;
 	return os;
 }
